Flaky: TestControllerGameServersNodeState
This involves two things:

(1) Adding a newFakeControllerWithMock to the metric test utilities to
allow for pre-population -- which with the client-go updates make it far
easier to write deterministic tests.

(2) Updated TestControllerGameServersNodeState with the new mock
construction.

This included removing the c.run() command, which can run c.collect() at
any point in time, and any number of times before the tests being
checked, and replaced this a c.collect() so we can control exactly when
metric collection occurs during the tests.