Expand feature freeze details during RC.
There have been a few discussions on what PRs get merged during
release candidate, which I think is down to the release documentation
not being as clear as it could be.

This should make the guidelines for what should go through and what
should be held back during release candidate clearer.