Add more more retries to htmltest
Either we have way more links or Github is being more aggressive with
it's 429 response to HTTP requests to check links.

So adding lots more retries to this test, to account for this.

We could potentially apply something like:
https://medium.com/@davidstanke/make-a-cloud-build-step-that-always-succeeds-9d23290a2f4e
long term if we want to always cache the htmltest link results even on
failure of this step.