Fix memory leak in client-go/workqueue
Since the Agones state is constantly changing, we got hit by this
memory leak in client-go rather badly:
https://github.com/kubernetes/kubernetes/issues/83895

A fix has been implemented, but is only available in 1.17.x client-go:
https://github.com/kubernetes/kubernetes/pull/83945

This PR does a backport of the fix into our `vendor` directly. Since in
the 1.11.0 release we'll be upgrading to support Kubernetes 1.17.x I
didn't do anything to ensure the fix would remain in place long term.