Updates to the release template
This PR includes two things:

1. The old helm install instructions worked for helm 2, but won't work
for helm 3 -- so made a small tweak to fix that (Closes #1886).
2. Switch the external links in the release template to use HTML over
markdown. This is because when copying the release from the release
draft in Github to the site, I always have to convert this by hand, so
that th html link checker ignores these links. This just makes things
the same on both, so that the release author doesn't have to manually
make the change.