Move supported site K8s version to shortcodes
Implements two new shortcodes:
* {{% k8s-version %}} - which outputs the currently supported version
* {{% k8s-api href="#podtemplatespec-v1-core" %}} - which outputs the
api reference url to the supported k8s version.

These shortcode utilise the `HUGO_ENV` environment to determine if it
should show the current K8s version, or the next - as it is only set to
the value of "production" when it is generated for the release version
of the agones.dev website.

Also added updates to the release checklist to manage this as well.

Ideally, this will remove lots of busy work of feature shortcoding a lot
of content as we churn through Kubernetes versions.

Long term, we may want to expand this to include separate tools for
current and next full semver versions - i.e. rather than just
1.16 ➡ 1.17, but something like 1.16.3 ➡ 1.17.14.

We cheat a little on this release as there is 1.16.13 and 1.17.13
released on both 1.16 and 1.17.

Work on #1824