Remove K8s API reference links from yaml examples
This PR removes the K8s API reference from the yaml files and redirects
users to look at the website reference, which has links to those API
docs.

This is to reduce toil as we increment supported Kubernetes version
numbers, as it is now far easier to edit the supported K8s version
throughout the entire website docs through config variables.

Work on #1824