nullable lastScaleTime on FleetAutoScaler CRD
lastScaleTime can be null as it is a pointer in the CRD definition.

Needs to be nullable in the CRD spec, as kubectl will pass this value as
"null", which causes the validation to fail in the error logs of the
controller (no functionality differences though).

Interestingly, creating a FleetAutoscaler from client-go code doesn't
cause this error, just kubectl.

Closes #1951