CRD OpenAPI Spec for ObjectMeta & PodTemplateSpec
This PR provides the mechanisms to pull core Kubernetes resource OpenAPI
schemas out of the API, and customise and convert them into a format
that is embeddable in our own CRD specifications.

This has been used to extract OpenAPI schemas for `ObjectMeta`, and
`PodTemplateSpec`, and then embedded in `GameServer` as well as parent
CRDs, creating a layer of validation for Pods definitions that are
implemented via a GameServer.

Also included in a new helm config option of
`gameservers.podPreserveUnknownFields` in case a user needs to escape
the field pruning on a Pod, or if PR causes a bug of any kind.

Also worth noting, this is not comprehensive validation, and therefore I
have not closed the the referenced issue.

Work on #1298