Terraform helm module fix deprecated version declaration (#1994)
This switches the helm terraform provider over to a required_providers block rather than the deprecated method of putting a version into the provider block. This should be done in the AKS/EKS modules too. Also adds a minimum version to terraform (this is in line with other modules). The source line in the block will cause 0.12 and lower to carp, but it will ignore the extra line. 0.13 and above will use it appropriately.

Co-authored-by: Mark Mandel <markmandel@google.com>