Fix bug in webhook docs after example switch
I didn't realise the webhook examples weren't under the
./examples/simple-game-server folder, and as such weren't already
updated for the 1.12.0 release branch.

So this broke the examples for webhooks, as they still pointed at the
`simple-udp` Fleet.

Doing a temporary redirect of the yaml examples to `main` for these
quickstart examples until the next release to resolve the issue.