Return last result even if all multicluster allocations fail (#2012)
Fixes #2011

This resolves an edge case where if all connections in a multicluster
allocation failed, the allocation would return (nil, nil) and cause
protobuf to fail to serialize the response. Instead, the last result
(even if it was unsuccessful!) is returned. An example of this is if
all clusters being fully allocated.