Release 1.13.0-rc
This includes all the changes for the 1.13.0 release candidate, as well
as several improvements to the release checklist and templates.

One thing of note, is that `simple-game-server:0.1` had moved to
`simple-game-server:0.2`, but hadn't been updated. This PR also includes
that change, as the code change was minor to the example.