Allocation: Drop stale GameServer on conflict
Previously, if an attempt at allocation would fail, it would get placed
automatically back into the local cache.

If the failure was because of a Conflict, there is no point in putting
it back into the cache, as it will never succeed until it gets updated
from the Kubernetes API.

Running at high load (TESTRUNSCOUNT=1 ./runAllocation.sh 200 20),
reduced the number of failed allocations due to contention from 298 down
to 180.

Work on #1856