Upgrade Rust language test version
Unblocking CI.

Looks like socket2 is using some newer language versions which fail to
compile on older versions of Rust.

Upgrading to the latest version of Rust to unblock.

If users want to keep the older version of socket2, it's a transitive
dependency, so they can specify it in their own Cargo.toml.