Cleanup: Start ➡ Run for all components. (#2058)
Minor cleanup that was irking me for a while. 😍

Almost all of the Agones component follow a pattern of `New*` for
construction and call `Run` to start a process / background thread /
etc.

Except for a couple of allocation components that have a `Start` instead
of a `Run`.

This PR renames their `Start` functions to `Run` to make everything
consistent.