Return HTTP 201 on GameServerAllocation
The standard for CRD creation is to return a HTTP 201 response. The Go
client is more flexible and was happy with a 200 response, but the C#
client will error out if anything but a 201 response.

This fixes things so we send the correct HTTP response code when
creating a GameServerAllocation.

Closes #2108