Refactor ReadyGameServerCache to AllocationCache
A bit of a chunky PR, but it does several things that were tied
together.

1. Since the cache will need to track both Ready and Allocated
GameServers, renamed everything to be "Allocation" related, rather than
just be about Ready GameServers
2. Implement functionality to be able to cache either Ready (for the
stable installation), or Allocated + Ready Gameservers for when the
Feature Flag of FeatureStateAllocationFilter is enabled.
3. There were some functionality and tests in some strange places, so
took the opportunity to clean that up, since I had to refactor a bunch
of stuff anyway.
4. Made sure tests would continue to pass in this interim state.

Work on #1239