GSA: Switch LabelSelector to GameServerSelector
This embeds the LabelSelector into a GameServerSelector struct, along
with the new player and state selector features we need for advanced
allocation, and utilising it in the `preferred` and `required` sections
of a `GameServerAllocation`.

This means the CRD doesn't change, but the Go api surface does.

This also moves the logic of validation and selection matching into the
GameServerSelector so it can be more easily reused.

GameServerAllocations are used pretty much everywhere, so a good chunk
of the PR is making that api change.

Work on #1239