Fix Rust Sample Docker image
Looks like the Docker image and accompanying build scripts weren't
updated with the move to Tonic, so it all went 💥.

This fixes that, removing lots of dependencies we no long need (yay
Tonic), upgrading the Rust version, and making sure all the relevant
code (e.g. protos) are copies to the appropriate place when creating the
Docker build image.

Work on #2179