Update proto and allocator for advanced allocation
Update the proto for allocation and related converters to allow advanced
allocation to also occur through the Allocation endpoint, and not just
the `GameServerAllocation` CRD.

This has been tested for backward compatability with the previous proto
version (ran allocation e2e from `main` branch against an install of
this PR).

The only outstanding item for #2148 is updates to the documentation!

Work on #2148