Flaky TestGameServerUnhealthyAfterReadyCrash
I noticed in a few flaky end-to-end tests, I kept seeing this in the
logs:

```
time="2021-10-08 19:02:01.419" level=info msg="sent UDP packet" address="35.247.94.25:7682" test=TestGameServerUnhealthyAfterReadyCrash
```

Over and over again, and also noting that it was happening _after_ the
e2e test had completed.

See:
https://console.cloud.google.com/cloud-build/builds/9ca5715a-443c-4693-bbd5-2879e61f2aaa;step=21?project=agones-images
https://console.cloud.google.com/cloud-build/builds/84cb8db2-1a11-4db3-a9e1-d9d51b9baf14;step=21?project=agones-images

My theory: The go routine had nothing in it that forced it to stop once
the test was complete - so depending on order of tests, it might keep
continuing for a while, while other tests ran.

If it did that, and a GameServer spun up on the same node and port as
the originally crashed GameServer, it would crash it - likely breaking
whatever test it ran into!

Work on #2296