Fix for health check race condition.
The Pod and GameServer data in the Informer cache are
eventually consistent and can result in data races.

This fix ensures that they are in sync at the time that
the GameServer is moved to Ready, so that Health checking
can behave correctly if a crash occurs after the fact.

This also includes a combination of smaller fixes/debug for e2e flakiness,
as we narrowed down the problem.

* Retries on UDP packet sending
* Failing on UDP packet sending dumps gameserver details
* Drop parallelism on e2e tests from 32 to 16.