Fleet Autoscaler custom sync: Race condition fix
Re-architected how the polling and autoscaling works such that it's more
inline with the original implementation:

* Switch from sync.Map to using a mutex, as it allows the code to be
transactional when doing multiple operations on the fasThread map.
* Switch out using the terminate channel into using a cancellable
context. This removes the possibility of deadlocks on an already closed
channel (issue in previous implementation).
* Move the sync loop to use `wait.Until()`, since it's already tested
and utilised.
* Make the loop only send an enqueue message rather than fire the
autoscale logic. This gives us back retry logic on failure, and also
means there are no errors that can be returned in that process that
could abort the syncing routing (which was in the previous
implementation).
* Moved to using the ObjectMeta.UID rather than name for the map key, as
we have no guarantee that the name is unique between runs (and some e2e
tests reused the same name, and caused race conditions).

Also includes some cleanup and extra debugging in tests as I worked
through what the originating issue was.

Closes #2385