Fix connection timeout on Rust SDK.
This is a fix for the Rust SDK flakiness!

The previous iteration of the code would fail as soon as the initial
connection would fail at any timeout greater than 2 seconds.

We move the connection to being lazy, so that the connection is only
attempted on first attempt, so we can retry connecting if there is a
failure for up to 30 seconds.