Added Agones Category to all UPROPERTY macro (#2587)
Unreal currently requires all UPROPERTY to have a Category.

This change adds the "Agones" category to all UPROPERTY's exposed in
Classes.h. It allows the AgonesSDK to be added to as an Engine Plugin or
Project Plugin. The motivation behind this was so Custom Online Subsystems
can reference the Agones code properly.