Fix local-includes WITH_ARM64=0 or WITH_WINDOWS=0
These variables are above where the local includes folder gets
imported, so they were not affected by any values that are stored in
there.

I have

```
# Just Linux
WITH_WINDOWS=0
WITH_ARM64=0
```

Within a dev.mk in that folder to ensure I only build and push linux
images, and that broke pushing, as with this dynamic variables being set
above the local include section, the `Makefile` attempts to also push
the arm64 image as well.