Install docs: Supported Container Architectures
Added a section in the installation docs where we specify the
supported cluster requirements. Now we're supporting multiple OS's and
architectures at different levels, I've added a table to indicate each
of the levels, so that users are aware.

I didn't keep arm64 for the past release, since we know the controller
is broken for 1.23.0.

As part of this work, I moved the note on node pools into its own
"best practices" section.

Work on #2216