Move CI image builds to Artifact Registry
Have removed the cleanup script for gcr.io, since it was overreaching on
what it should delete -- so part of that work is to move CI builds to
Artifact Registry, where we can keep production images separate from CI
builds.

Have created us-docker.pkg.dev/agones-images/ci for images for and
related to CI. These changes should make it so that Agones PR
image builds to push to that registry, and testing should happen with
those images.

Production images will stay the same for now, this is just a first step.

Work on #2358