Fix minikube dev tooling (#2662)
The update to multi-arch images broke the minikube dev tooling, this
fixes that.

Only caveat being, that I can only test on amd64 architecture. In
theory, it might work on a M1, but if not, we can switch our which
images are loaded before being re-tagged based on host architecture.