Add Feature Flag section to troubleshooting
On Slack, I've found that I've had to explain how to check if Feature
Flags are set correctly or not - so writing it down so it's easier to
find.

We do validation within parsing feature flags, but there are several
layers in which one could mess it up (could be setting `true` when
meaning `false`), so good to have some documentation.