Fleet scale down: Remove < Ready GameServers first
When scaling down GameServers within a single node, Agones would remove
GameServers that would be there the longest.

This would mean that what would usually happen is that Ready
GameServers would be shut down preferentially over GameServers that
where not yet Ready.

This would result in churn on the node, and a drop in Ready GameServers
until the newer ones came to a Ready state.

This change will prefer GameServers that are in a state before Ready for
deletion, when comparing GameServers that are stored on a single node.

Closes #2372