TestGameServerRestartBeforeReadyCrash: Close race
This closes the race in #2445 by introducing a larger delay before
we mark the game server pod as Ready(). This change admits the
possibility that in some circumstances, if the game server initializes
too quickly and kubelet loses a race to update the pod, we may perceive
the game server as having crashed when it did not.

Closes #2445.