fix: solve the native collection's memory leak detected by Unity 2021.3.7f1
On Unity 2021.3.7f1, several error messages says "A Native Collection has not been disposed, resulting in a memory leak, Allocated from:xxx"; When look into this issue, it points into AgonesSdk and the UnityWebRequest's Dispose has not been called. After calling the Dispose() at the end of the method, the same error disappeared.