LifecycleContract alpha: Implement (immutable) scale subresource, add pdb (#2807)
This PR, under the `LifecycleContract` feature gate suggested in #2794

* Adds a scale resource to `GameServer` by adding an
`immutableReplicas` field, which has a `default`, `min` and `max` of 1.
Having a `scale` subresource lets us define a `PodDisruptionBudget`
that can be set to `maxUnavailable: 0%` [1].

* Adds a PDB per namespace with label selector
`agones.dev/safe-to-evict: "false"`, which nothing yet adds.

* Adds a mechanism to get feature gate values in Helm by using:
  {{- $featureGates := include "agones.featureGates" . | fromYaml }}

* Cleanup / documentation of feature gate mechanisms

After this PR, it's possible to define a fleet with the label and have
all `GameServer` pods protected by a `PodDisruptionBudget`, e.g.:

```
$ kubectl scale fleet/fleet-example --replicas=5
fleet.agones.dev/fleet-example scaled
$ kubectl describe pdb
Name:             agones-gameserver-safe-to-evict-false
Namespace:        default
Max unavailable:  0%
Selector:         agones.dev/safe-to-evict=false
Status:
    Allowed disruptions:  0
    Current:              4
    Desired:              5
    Total:                5
Events:                   <none>
```

Additionally, because min/max/default are 1, Kubernetes enforces the
immutability for us:

```
$ kubectl scale gs/fleet-example-k6dfs-6m5nq --replicas=1
gameserver.agones.dev/fleet-example-k6dfs-6m5nq scaled
$ kubectl scale gs/fleet-example-k6dfs-6m5nq --replicas=2
The GameServer "fleet-example-k6dfs-6m5nq" is invalid: spec.immutableReplicas: Invalid value: 2: spec.immutableReplicas in body should be less than or equal to 1
$ kubectl scale gs/fleet-example-k6dfs-6m5nq --replicas=0
The GameServer "fleet-example-k6dfs-6m5nq" is invalid: spec.immutableReplicas: Invalid value: 0: spec.immutableReplicas in body should be greater than or equal to 1
```

The only artifact of this addition is a new field in the Spec/Status
named `immutableReplicas`, in the Kubernetes object. This field is not
present in the in-memory representation for `GameServer`, nor is it
present in `etcd` (by defaulting rules). The field is visible on
`describe` or `get -oyaml`, but is otherwise ignored.

[1] https://kubernetes.io/docs/tasks/run-application/configure-pdb/#identify-an-application-to-protect