cloudproduct: Register API hooks, move validation/mutation to API
This commit fixes the fact that currently cloud product specific API
validation/mutation is in a weird state where it has to occur outside
the pkg/apis/agones layer and instead currently lives in the
controller. The mechanical reason for this previously was to avoid an
import cycle - pkg/apis/agones/v1 could not import cloudproduct
because cloudproduct needs the v1 type definitions.

Instead of shoving everything under one interface, I split
CloudProduct up into two interfaces, the "controller hooks" (new port
allocator, sync ports back) and the "api hooks". In the process, I
ahead and converted the cloud product layer into a global singleton
"initialize once" layer so all registration happens together. Now when
we initialize, the controller hooks are saved as a singleton, and the
API hooks are registered with the API layer. This last step breaks the
import cycle, as the API layer can define its own hook interface and
the cloud products can provide that.

I then adopted the new pattern in the API layer and moved the logics
out of the controller layer. This fixes a latent bug whereby the Fleet
and GameServerSet controllers are not currently doing any cloud
product specific GameServerSpec validation.

Along the way, I also changed `ValidateGameServer` to
`ValidateGameServerSpec` and `MutateGameServerPod` to
`MutateGameServerPodSpec` to narrow access in both to the `Spec`
rather than the whole object.