lint: Reenable `revive`
The way revive was configured, we configured the confidence interval,
but didn't enable any checks, it seems. By deleting the message, it
reverts to the default `revive` config, which includes:

* Comments on exported symbols must be of the right form (this was
noticed as lacking in #2868)

* Package comments - I'm guessing from the number of changes that
we've never had this linter, but given that certain IDEs expect the
package to be offset from the header, having this linter enforces it.

Closes #2873