Inline JSON: GameServerAllocation v1.LabelSelector
The K8s YAML library will inline embeded structs, but some other ones
will not. So adding the `json:",inline"` hint, so that they know that
the `v1.LabelSelector` should be inlined if people are so inclined to
reference it via Golang yaml libraries.

Include regeneration of the crd docs and client libraries, which
picked up on some small changes.

Closes #2853