SafeToEvict: Implement Eviction API, add SetEviction cloud product hook (#2857)
* SafeToEvict feature: Implement the `eviction.safe` API logic

Towards #2794: This commit implements the defaulting/validation for
the eviction.safe field, under the SafeToEvict feature gate. In
particular, it enshrines the defaulting logic for the somewhat
complicated backwards compatibility case (safe-to-evict=true is set
but SafeToEvict is not).

Along the way, refactor a very repetitive unit test table.

* Ran gen-api-docs

* Implement SafeToEvict enforcement of safe-to-evict annotations/label

This commit adds the interpretation of GameServer.Status.Eviction:

* Following the pattern in #2840, moves the enforcement of
cloudproduct specific things up to the controller, adding a new cloudproduct
SetEviction hook to handle it.
  * As a result, when the SafeToEvict feature flag is on, we disable annotation
    in gameserver.go and test that we don't set anything (as a change detection test).

* Adds a generic and GKE Autopilot version of how to set policy from the
eviction.safe status. In Autopilot, we skip using safe-to-evict=false.