Don't run GameServerSpec validation on development GameServers (#2889)
Development GameServers are basically just metadata placeholders and
need no cloud-product-specific validation. This fixes the
TestDevelopmentGameServerLifecycle e2e on Autopilot.

Along the way: Fix TestFleetGSSpecValidation e2e on Autopilot. The
intent of the last test is a success case, which we can achieve more
simply than using a Static port. Leave the rest to unit tests.