e2e framework: Timing based on cloud product, end on terminal states
Most of the "wait for state" functions use a hardcoded 5m. To allow
for autoscaling variance on Autopilot, allow for 10m on Autopilot.

In addition, the WaitForGameServerState function can understand which
states are terminal to save some time in the case that a GameServer
has already moved into a terminal state.

Along the way: Remove New/NewWithRates as they are unused or
unnecessary wrappers - in particular, they definitely don't need to be
exported.

Towards #2777