controller/extensions: Add explicit requests/limits for ephemeral storage
On Autopilot, ephemeral storage defaults to a request/limit of 1GiB,
but on Agones we assume there's no ephemeral storage limit. The
default persistent log limit is 10GB.

In this PR, we adds a default resources.{limits,requests} block to the
controller if the user has not already specified `resources` and set
it to the persistent log maximum plus a fudge factor of 100MB to allow
for the certs and some overhead. By default this comes out to 10100MB,
which slides under the Autopilot 10GiB limit by virtue of GiB/MB
differences.

Towards #2777