e2e: Use gotestsum for CI e2e runs, change e2e ARGS handling (#2904)
The main goal of this PR is to add gotestsum
(https://pkg.go.dev/gotest.tools/gotestsum#section-readme) as an
optional way to run e2es. gotestsum wraps `go test -json` and, aside
from some fancy formatting options, allows us to re-run flaky tests.
As we start adding more permutations to the test matrix, this will
become important. (Also important will become actually monitoring
flakes, but that's the next step.)

We do this by:
* Installing `gotestsum` in the build image
* Adding a new E2E_USE_GOTESTSUM option to the Makefile that defaults
to 2x reruns and super-terse output.
* Also added a GOTESTSUM_TESTNAME_AND_VERBOSE option meant for CI
that shows the status of all of the tests then dumps the full verbose
output for diagnostics.

One thing that changes slightly after this PR is the arguments
involved in invoking e2es via `make`. Previously, both `ARGS` and
`GO_E2E_TEST_ARGS` were passed to `go test`. In order to hack
arguments for re-runs, `gotestsum` requires a delineation between the
arguments to `go test` and the arguments to the test binary. However,
so that usage is consistent when E2E_USE_GOTESTSUM is present or not,
I changed both to do the equivalent of:

```
go test ${ARGS} -args ${GO_E2E_TEST_ARGS}
```

In other words, for the e2e targets:
* Use `ARGS` to pass args to `go test` (e.g. `-parallel 16`)
* Use `GO_E2E_TEST_ARGS` to pass args to the test binary (e.g. `--kubeconfig`)
* If `E2E_USE_GOTESTSUM` is on, use `GOTESTSUM_ARGS` to pass args to `gotestsum` (e.g. `--rerun-fails=5`)