Validate the .scheduling field in the fleet and gameserverset (#2892)
Adds a cloud product hook specific to the scheduling field, for use in
the fleet / gameserverset APIs. This might seem narrowly scoped, but
this field is the only one that propagates down outside the template.

Modifies one e2e to test that Autopilot rejects appropriately. Adds
several skips to other tests.

Towards #2777