Allow 30m more to acquire consul lock (#2922)
We seem to be having more frequent consul lock timeouts on CI. Best
guess, this is a bandwidth issue of all PRs using the same e2e
clusters and a higher number of PRs, plus probably some organic growth
in e2e length. This PR is a stop-gap:

* Bumps the consul lock timeout from 1h to 1.5h - basically "wait our
turn" for longer.

* Bumps the whole build timeout from 2h to 2.5h. I assume the whole
build has a generous timeout to allow for the image build, etc, prior
to e2es, so kept the increase here 1:1.

* Bumps the queue TTL from (default) 1h to 6h. This is a generous bump
to allow builds to hang out longer in the queue.