Bump the cloud build expiration time, we can no longer retry (#2927)
Along the way: I was trying to audit uses of FeatureTestMutex and
noticed that the user of EnableAllFeatures doesn't bother. I think
this part of the change is a no-op as these tests aren't marked
t.Parallel(), but changed it for consistency in case anyone else does
a hand audit.