Queue running e2es in fair order
This commit changes the build steps involved in running e2es so that
prior to running e2es, the current build is the oldest running build
within the same trigger (the agones branch trigger, in our case).

This effectively turns the e2e-wait-to-become-leader step into a
custom queuing mechanism. The idea is that all other steps of the
build can run in parallel, but only the e2e leader can run e2es, and
the e2e leader is always the oldest build in the queue.

I also changed the second test in each config (e2e-stable*) to depends
on the first (e2e-feature-gates*). So now the build looks roughly
like:

become-leader +---> feature-gates-config-1 ----> stable-config-1
               \--> feature-gates-config-2 ----> stable-config-2

(where config-1 and config-2 are standard and autopilot right now).

By serializing the builds then the steps under it, we can remove the
use of consul. I kept consul in for now because until this commit is
present in every PR, other PRs won't respect the same ordering (since
the cloudbuild.yaml is pulled from branch).

See the PR for more details on the problem this is fixing.