Disable consul locking if consul is not present
This is a continuation of #2932: The idea here is two-fold:

* Make the use of consul locking conditional on whether consul is
present. This lets us stop and/or restart using consul locking
out-of-band of a PR, which can prevent us from getting in a funny
position of needing to force merge because CI is broken.

* Uninstalling consul has a certain finality to it: It's not uncommon
in OSS projects to see PRs that are _very_ out of date. If we were to
just stop using consul but leave it installed, those PRs would run
through the old locking mechanism unimpeded and likely harm an
in-progress e2e. If we instead uninstall consul, stale PRs will just
fail CI - a rebase will then pick up the new locking mechanism.

I will test this on the Autopilot CI first, since it's still in "allow
failure" mode, making it a perfect canary.