Remove apiHooks singleton
Apologies for the size of this - I iterated like 3 or 4 different ways
to do this and, believe it or not, this was the least painful:

Instead of using a singleton `apiHooks`, this PR changes all things
that depend eventually on apiHooks to take explicit interfaces. In
most cases, that's the v1.APIHooks interface, which stays, but in a
couple of cases I minimized to the specific portion needed. We do this
by:

* Reverting a lot of the nature of #2868, but keeping the interface
"layering" with a chunk of the interface defined in agonesv1/ and the
remainder outside.

* Plumbing hooks from the top as before.

* Introduces a "FakeAPIHooks" interface for unit tests to use.
Crucially, this interface is not at all dependent on feature gates.
  * As a call out, the fakes don't quite last all the way to
    to the controller tests that create a full controller. Those
    use the generic provider. I may poke at this more to remove
    the influence of feature gates here as well.

Towards #2926, but I'm not calling it fixed with this - I think we
should bust the feature gates global in the runtime package as well to
make the unit tests more predictable, but this ~mostly resolves the
`apiHooks` cases.