Update the apiserver request dashboard
This is nearly unreviewable, so I'm attaching a screenshot to the PR.

Changes:

* The original metrics were deprecated. `apiserver_request_total`
exactly replaces `apiserver_request_count`, but the old
`apiserver_request_latencies_summary`, which was a gauge, was replaced
by `apiserver_request_duration_seconds_bucket`, which is a historgram
type. Changed the expressions to match these changes.

* Transitioned to $__rate_interval instead of 5m, per recommendation:
https://grafana.com/docs/grafana/latest/datasources/prometheus/template-variables/#use-__rate_interval

* Fixed an issue where `label_values(label)` isn't supported on GMP (it
seems - probably because this goes across ALL metrics), but
`label_values(label, metric)` is - this is a change I'll need to make
on nearly every other dashboard but I'll get there later.

* Upgraded the graph from a deprecated graph type to the new time
series type.