Fix for `make gen-crd-client`
Having `agonesv1.FakeApiHooks` be public means that the code generation
toolkit will see it as a generation target, and then get all confused.

Fix is to make it private, and keep it's stubbable capabilities.

Since we also need a mock for other parts of the Agones system,
I also created a no-op version for general usage as well in the
`testing` package. (We can't use this one in `agonesvs`, since Go
will complain about cyclic dependencies).